Migrate relevant test cases
