+++ lib/esbonio-extensions/tests/data/tutorial/image.rst
